Added test for build script
I added some initial tests for verifying the outputs of the db build script. A few issues were identified however, we need to pin the blast version a user can use as outputs between blast versions seem to vary quite a bit, the test output data was updated to meet blast+ version 2.15.0 and the readme's compatibility section was updated. One test currently fails as well, as there is a lack of continuity between outputs in the results.json and the config.json particularly the date fields.

It would be nice to make the keys between these two files match in the future as well.